requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.4
django-ipware==4.0.2
django-sesame==2.4
psycopg2==2.9.3
sqlparse==0.4.2
